pyproject.toml
[tool.poetry]
name = "kaggle-isic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
lightgbm = "^4.5.0"
pyarrow = "^17.0.0"
torch = "^2.4.0"
h5py = "^3.11.0"
pillow = "^10.4.0"
albumentations = "^1.4.14"
timm = "^1.0.9"
torcheval = "^0.0.7"
notebook = "^7.2.2"
catboost = "^1.2.5"
pytorch-tabnet = "^4.1.0"
optuna = "^3.6.1"
kaggle = "^1.6.17"
numpy = "1.26.4"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
